Adding item glow effect

+++ src/main/java/me/otho/customItems/configuration/jsonReaders/common/Cfg_drop.java
@@ -5,5 +5,5 @@ public class Cfg_drop {
 	public String id;	
 	public int min = 1;
     public int max = 1;    
-    public int chance = 50;
+    public float chance = 50.0f;
 }
